# src/api/swagger/v1.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Room": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"places": {
						"type": "number",
						"format": "double"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"hotel": {
						"$ref": "#/components/schemas/Hotel",
						"description": "SET datatype is not supported by SQLite:"
					},
					"isActive": {
						"type": "boolean"
					},
					"reservations": {
						"items": {
							"$ref": "#/components/schemas/Reservation"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"places",
					"price",
					"hotel",
					"isActive",
					"reservations"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Hotel": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"postcode": {
						"type": "string"
					},
					"rooms": {
						"items": {
							"$ref": "#/components/schemas/Room"
						},
						"type": "array",
						"description": "SET datatype is not supported by SQLite:"
					}
				},
				"required": [
					"id",
					"name",
					"address",
					"postcode",
					"rooms"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReservaionStatus": {
				"enum": [
					"CREATED",
					"PAID",
					"CONFIRMED",
					"CANCELLED",
					"FULFILLED"
				],
				"type": "string"
			},
			"Reservation": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/ReservaionStatus"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"room": {
						"$ref": "#/components/schemas/Room"
					},
					"fromDate": {
						"type": "string",
						"format": "date-time"
					},
					"toDate": {
						"type": "string",
						"format": "date-time"
					},
					"places": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"status",
					"created",
					"user",
					"room",
					"fromDate",
					"toDate",
					"places"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"authKey": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"reservations": {
						"items": {
							"$ref": "#/components/schemas/Reservation"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"firstName",
					"lastName",
					"authKey",
					"isActive",
					"reservations"
				],
				"type": "object",
				"additionalProperties": false
			},
			"roomFindResponse": {
				"properties": {
					"rooms": {
						"items": {
							"$ref": "#/components/schemas/Room"
						},
						"type": "array"
					}
				},
				"required": [
					"rooms"
				],
				"type": "object",
				"additionalProperties": false
			},
			"roomFindRequest": {
				"properties": {
					"from": {
						"type": "string",
						"format": "date-time"
					},
					"to": {
						"type": "string",
						"format": "date-time"
					},
					"places": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"from",
					"to",
					"places"
				],
				"type": "object",
				"additionalProperties": false
			},
			"reservationMakeResponse": {
				"properties": {
					"result": {
						"$ref": "#/components/schemas/Reservation"
					}
				},
				"required": [
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"reservationMakeRequest": {
				"properties": {
					"from": {
						"type": "string",
						"format": "date-time"
					},
					"to": {
						"type": "string",
						"format": "date-time"
					},
					"places": {
						"type": "number",
						"format": "double"
					},
					"roomId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"from",
					"to",
					"places",
					"roomId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"reservationChangeResponse": {
				"properties": {
					"result": {
						"$ref": "#/components/schemas/Reservation"
					}
				},
				"required": [
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"reservationChangeRequest": {
				"properties": {
					"from": {
						"type": "string",
						"format": "date-time"
					},
					"to": {
						"type": "string",
						"format": "date-time"
					},
					"places": {
						"type": "number",
						"format": "double"
					},
					"reservationId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"from",
					"to",
					"places",
					"reservationId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"reservationCancelResponse": {
				"properties": {
					"result": {
						"$ref": "#/components/schemas/Reservation"
					}
				},
				"required": [
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"reservationCancelRequest": {
				"properties": {
					"reservationId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"reservationId"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"apiKey": {
				"type": "apiKey",
				"name": "X-API-KEY",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "booking-test",
		"version": "1.0.0",
		"description": "Booking test",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Sergiy Savostin"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/v1/room/find": {
			"get": {
				"operationId": "Find",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/roomFindResponse"
								}
							}
						}
					}
				},
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "from",
						"required": true,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "to",
						"required": true,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "places",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/reservation/make": {
			"put": {
				"operationId": "Make",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/reservationMakeResponse"
										},
										{
											"type": "boolean",
											"enum": [
												false
											]
										}
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/reservationMakeRequest"
							}
						}
					}
				}
			}
		},
		"/api/v1/reservation/change": {
			"post": {
				"operationId": "Change",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/reservationChangeResponse"
										},
										{
											"type": "boolean",
											"enum": [
												false
											]
										}
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/reservationChangeRequest"
							}
						}
					}
				}
			}
		},
		"/api/v1/reservation/cancel": {
			"post": {
				"operationId": "Cancel",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/reservationCancelResponse"
										},
										{
											"type": "boolean",
											"enum": [
												false
											]
										}
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/reservationCancelRequest"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}